Updated OffersCommand.
